refactor(hub): relocate hub coordinator modules to aether-hub crate (ADR-0071 phase 7d-2)
Mechanical relocation: moves the engine TCP listener, MCP transport,
session/registry/spawn/log_store stores, loopback substrate, chassis,
and `HubServerCapability` driver from the `aether-substrate-hub`
binary crate into the `aether-hub` library crate, per the ADR-0070
phase 5 directory sketch:

```
aether-hub/                           — owns hub-shaped capabilities
├── HubClientCapability (passive)     — already shipped (phase 7a/b)
└── HubServerCapability (driver)      — moved here this PR

aether-substrate-hub/                 — thin main, delegates to aether-hub
└── main.rs                           — env reads + HubChassis::build(env)?.run()
```

* `git mv` moves chassis.rs, engine.rs, log_store.rs, loopback.rs,
  mcp.rs + mcp/, registry.rs, session.rs, spawn.rs from
  aether-substrate-hub/src/ to aether-hub/src/. Integration tests
  (engine_handshake.rs, engine_logs.rs) follow into
  aether-hub/tests/.
* `aether-hub/src/lib.rs` declares the new modules + re-exports
  every type that was reachable through `aether_substrate_hub::*`
  before. Adds `run_engine_listener` and `DEFAULT_ENGINE_PORT` (the
  two top-level items the old aether-substrate-hub lib carried).
* `aether-hub/Cargo.toml` absorbs the deps the moved modules
  use: axum, base64, bytemuck, postcard, rmcp (server + macros +
  http-streamable transport), schemars, serde_json, tokio (with
  signal/process/fs features), tokio-util, uuid, libc (cfg(unix)).
* `aether-substrate-hub/Cargo.toml` shrinks to `aether-hub` plus
  `wasmtime` (for `wasmtime::Result` in main). The `[lib]` section
  goes away — the crate is binary-only now.
* `aether-substrate-hub/src/main.rs` rewires its imports from
  `aether_substrate_hub::` to `aether_hub::`. Tests' imports get
  the same rename.

No behavior change. The hub coordinator, child cleanup, signal
handling, MCP tool surface, session routing, log store — all
unchanged.

Closes ADR-0070 phase 5 / ADR-0071 phase 7's `aether-hub` crate
goal: the crate now owns both the `HubClientCapability` (passive
engine-side) and the `HubServerCapability` (driver hub-side).
