+++ crates/aether-hub/Cargo.toml
@@ -3,15 +3,46 @@ name = "aether-hub"
 version.workspace = true
 edition.workspace = true
 
-# ADR-0070 phase 4 / ADR-0071 phase 7: hub-protocol-aware components
-# (the substrate-side `HubClientCapability`, the engine-mailbox reader
-# / writer / heartbeat thread cluster, the future hub server) live
-# here so `aether-substrate-core` doesn't depend on hub framing for
-# chassis that don't bridge to a hub.
+# ADR-0070 phase 4-5 / ADR-0071 phase 7: hub-protocol-aware substrate
+# code lives here so `aether-substrate-core` doesn't depend on hub
+# framing for chassis that don't bridge to a hub. Holds:
+#
+# - the substrate-side hub client + capability (`HubClient`,
+#   `HubProtocolBackend`, `HubClientCapability`),
+# - the engine-facing TCP listener and per-connection read-loop,
+# - the rmcp-driven MCP transport + tool surface,
+# - the engine / session / spawn / log registries the hub coordinates,
+# - the `HubServerCapability` driver and `HubChassis` builder shape.
+#
+# The `aether-substrate-hub` binary crate consumes this and shrinks to
+# just `main.rs`.
 
 [dependencies]
 aether-data = { path = "../aether-data" }
 aether-hub-protocol = { path = "../aether-hub-protocol" }
+aether-codec = { path = "../aether-codec" }
+aether-kinds = { path = "../aether-kinds" }
 aether-substrate-core = { path = "../aether-substrate-core" }
+axum = "0.8"
+base64 = "0.22"
+bytemuck = { version = "1.19", features = ["derive"] }
+postcard = { version = "1.1", default-features = false, features = ["alloc"] }
+rmcp = { version = "0.8", features = [
+    "server",
+    "macros",
+    "transport-streamable-http-server",
+] }
+schemars = "1"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+tokio = { version = "1", features = ["rt-multi-thread", "macros", "net", "io-util", "sync", "time", "signal", "process", "fs"] }
+tokio-util = "0.7"
 tracing = "0.1"
+uuid = { version = "1", features = ["v4", "serde"] }
 wasmtime = "30"
+
+[target.'cfg(unix)'.dependencies]
+libc = "0.2"
+
+[dev-dependencies]
+tokio = { version = "1", features = ["rt-multi-thread", "macros", "net", "io-util", "sync", "time"] }
