+++ crates/aether-hub/src/mcp.rs
