+++ crates/aether-hub/src/mcp/args.rs
