+++ crates/aether-hub/src/mcp/codecs.rs
