+++ crates/aether-hub/src/mcp/tools.rs
