+++ crates/aether-hub/src/registry.rs
