+++ crates/aether-substrate-hub/Cargo.toml
@@ -3,32 +3,16 @@ name = "aether-substrate-hub"
 version.workspace = true
 edition.workspace = true
 
+# ADR-0071 phase 7d-2 collapsed this crate to a binary-only entry
+# point: every module the hub needs (the engine listener, MCP
+# transport, session/registry/spawn/log_store stores, loopback
+# substrate, chassis + driver capability) lives in `aether-hub`.
+# `main.rs` reads env vars and calls `HubChassis::build(env)?.run()`.
+
+[[bin]]
+name = "aether-substrate-hub"
+path = "src/main.rs"
+
 [dependencies]
-aether-hub-protocol = { path = "../aether-hub-protocol" }
-aether-kinds = { path = "../aether-kinds" }
-aether-data = { path = "../aether-data" }
-aether-codec = { path = "../aether-codec" }
-aether-substrate-core = { path = "../aether-substrate-core" }
 aether-hub = { path = "../aether-hub" }
-axum = "0.8"
-base64 = "0.22"
-bytemuck = { version = "1.19", features = ["derive"] }
-postcard = { version = "1.1", default-features = false, features = ["alloc"] }
-rmcp = { version = "0.8", features = [
-    "server",
-    "macros",
-    "transport-streamable-http-server",
-] }
-schemars = "1"
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-tokio = { version = "1", features = ["rt-multi-thread", "macros", "net", "io-util", "sync", "time", "signal", "process", "fs"] }
-tokio-util = "0.7"
-uuid = { version = "1", features = ["v4", "serde"] }
 wasmtime = "30"
-
-[target.'cfg(unix)'.dependencies]
-libc = "0.2"
-
-[dev-dependencies]
-tokio = { version = "1", features = ["rt-multi-thread", "macros", "net", "io-util", "sync", "time"] }
